move runValidatorPlan into its own file
